updated exercise 4 file structure
